requirements.txt
asgiref==3.3.4
atlassian-python-api==3.10.0
certifi==2021.5.30
chardet==4.0.0
Deprecated==1.2.12
Django==3.2.3
django-common-helpers==0.9.2
idna==2.10
nested-lookup==0.2.22
oauthlib==3.1.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
wrapt==1.12.1
